backend/requirements.txt
fastapi
uvicorn
authlib
python-dotenv
starlette
sentence-transformers
pinecone-client
langchain
redis
pdfplumber
transformers
torch
scikit-learn
numpy
pydantic
httpx
